popup show on delete order

+++ resources/views/pages/admin/tour/table.blade.php
@@ -67,7 +67,7 @@
                             {{ Icon::edit() }}
                         </a>
                         @if(!$tour->trashed())
-                            <a href="#" title="Delete" onclick="event.preventDefault();document.getElementById('tour-{{ $tour->id }}-delete').submit();" class="btn btn-outline-danger btn-sm mb-1">
+                            <a href="#" title="Delete" onclick="if(confirm('Are you sure you wish to cancel this order?')) { event.preventDefault();document.getElementById('tour-{{ $tour->id }}-delete').submit(); } else { return false; }" class="btn btn-outline-danger btn-sm mb-1">
                                 {{ Icon::delete() }}
                             </a>
                             <form id="tour-{{ $tour->id }}-delete" action="{{ route('tours.delete', ['tour' => $tour,]) }}" method="POST"
